server/pyproject.toml
[project]
name = "server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.18.4",
    "asyncpg>=0.31.0",
    "fastapi>=0.128.8",
    "openai>=2.20.0",
    "pydantic-settings>=2.11.0",
    "python-dotenv>=1.2.1",
    "sqlalchemy>=2.0.46",
    "uuid7>=0.1.0",
    "uvicorn[standard]>=0.39.0",
]

[dependency-groups]
dev = [
    "ruff>=0.15.1",
]
